backend/requirements.txt
fastapi
uvicorn
python-multipart
sentence-transformers
easyocr
faiss-cpu
numpy
requests
Pillow
python-dotenv
opencv-python-headless 
huggingface_hub         